Refactor: shared/ui 재사용성 컴포넌트 분리
