+++ src/shared/ui/index.ts
@@ -1,0 +1,7 @@
+export * from "./Button";
+export * from "./Card";
+export * from "./Dialog";
+export * from "./Input";
+export * from "./Select";
+export * from "./Table";
+export * from "./TextArea";
